fix: correct leaderboard avgRating type and improve packages UI
leaderboard:
- Cast avgRating to float in SQL query to avoid string return type
- Wrap avgRating with Number() in client to ensure .toFixed() works

packages-manager: refresh stats bar with card layout and icons
package-builder-form: update accent color to brand blue, styling tweaks

+++ app/(public)/leaderboard/page.tsx
@@ -38,7 +38,7 @@ export default async function LeaderboardPage() {
       totalOrders: editors.totalOrders,
       isAvailable: editors.isAvailable,
       avgRating: sql<number | null>`
-        (SELECT ROUND(AVG(r.rating)::numeric, 1)
+        (SELECT ROUND(AVG(r.rating)::numeric, 1)::float
          FROM reviews r
          INNER JOIN orders o ON r.order_id = o.id
          WHERE o.editor_id = ${editors.id} AND r.role = 'client')
